refactor: migrate all link API handlers to api/links/ module
Extract all link management handlers from router.rs into a dedicated
api/links/ module following the established layered architecture.

- Handlers now call LinkRepository directly instead of db:: wrappers
- Removed 16 now-dead forwarding functions from db/queries.rs
- Updated lib.rs, openapi.rs, api/admin/users.rs, and
  api/orgs/invitations.rs to reference new api::links paths
- get_frontend_url and sync_link_mapping_from_link re-exported from
  api/links for use by other modules
