+++ src/api/admin/users.rs
@@ -1,3 +1,4 @@
+use crate::api::links::sync_link_mapping_from_link;
 /// Admin user-management handlers
 ///
 /// GET    /api/admin/users           — list all users (paginated)
@@ -10,7 +11,6 @@ use crate::auth;
 use crate::kv;
 use crate::models::link::LinkStatus;
 use crate::repositories::UserRepository;
-use crate::router::sync_link_mapping_from_link;
 use crate::utils::AppError;
 use worker::d1::D1Database;
 use worker::*;
