--- src/api/links.rs
@@ -1,2 +1,0 @@
-// Link management API - Phase 4
-// Most functionality moved to router.rs
