+++ src/repositories/mod.rs
@@ -18,6 +18,7 @@ pub mod analytics_repository;
 pub mod api_key_repository;
 pub mod billing_repository;
 pub mod blacklist_repository;
+pub mod link_repository;
 pub mod org_repository;
 pub mod product_repository;
 pub mod report_repository;
@@ -28,6 +29,7 @@ pub use analytics_repository::AnalyticsRepository;
 pub use api_key_repository::ApiKeyRepository;
 pub use billing_repository::BillingRepository;
 pub use blacklist_repository::BlacklistRepository;
+pub use link_repository::LinkRepository;
 pub use org_repository::OrgRepository;
 pub use product_repository::ProductRepository;
 pub use report_repository::ReportRepository;
