+++ src/router.rs
@@ -1,46 +1,26 @@
 use crate::auth;
 use crate::db;
-use crate::kv;
 use crate::middleware::{RateLimitConfig, RateLimiter};
-use crate::models::{
-    PaginatedResponse, PaginationMeta, Tier,
-    link::{CreateLinkRequest, Link, LinkStatus, UpdateLinkRequest},
-};
-use crate::repositories::tag_repository::{TagRepository, validate_and_normalize_tags};
-use crate::utils::{generate_short_code, now_timestamp, validate_short_code, validate_url};
 use chrono::{Datelike, TimeZone};
-use std::future::Future;
-use std::pin::Pin;
 use worker::d1::D1Database;
 use worker::*;
 
-/// Get the frontend URL from environment, with localhost fallback for local dev
-pub fn get_frontend_url(env: &Env) -> String {
-    env.var("FRONTEND_URL")
-        .map(|v| v.to_string())
-        .unwrap_or_else(|_| "http://localhost:5173".to_string())
-}
-
 /// Extract client IP from Cloudflare headers
 fn get_client_ip(req: &Request) -> String {
-    // Try CF-Connecting-IP first (most reliable with Cloudflare)
     if let Ok(Some(ip)) = req.headers().get("CF-Connecting-IP") {
         return ip;
     }
-
     // Fallback to X-Forwarded-For
-    if let Ok(Some(forwarded)) = req.headers().get("X-Forwarded-For") {
+    if let Ok(Some(forwarded)) = req.headers().get("X-Forwarded-For")
+        && let Some(ip) = forwarded.split(',').next()
+    {
         // Take first IP in the list
-        if let Some(ip) = forwarded.split(',').next() {
-            return ip.trim().to_string();
-        }
+        return ip.trim().to_string();
     }
-
     // Fallback to X-Real-IP
     if let Ok(Some(ip)) = req.headers().get("X-Real-IP") {
         return ip;
     }
-
     // Last resort: use a placeholder (should never happen with Cloudflare)
     "unknown".to_string()
 }
@@ -52,1644 +32,6 @@ fn hash_ip(ip: &str) -> String {
     let mut hasher = DefaultHasher::new();
     ip.hash(&mut hasher);
     format!("{:x}", hasher.finish())
-}
-
-pub(crate) async fn resolved_forward_for_link(db: &D1Database, link: &Link) -> bool {
-    if let Some(value) = link.forward_query_params {
-        value
-    } else {
-        db::get_org_forward_query_params(db, &link.org_id)
-            .await
-            .unwrap_or(false)
-    }
-}
-
-pub(crate) async fn sync_link_mapping_from_link(
-    db: &D1Database,
-    kv_store: &worker::kv::KvStore,
-    link: &Link,
-) -> Result<()> {
-    match link.status {
-        LinkStatus::Blocked => {
-            kv::delete_link_mapping(kv_store, &link.org_id, &link.short_code).await
-        }
-        LinkStatus::Active | LinkStatus::Disabled => {
-            let resolved_forward = resolved_forward_for_link(db, link).await;
-            let mapping = link.to_mapping(resolved_forward);
-            kv::update_link_mapping(kv_store, &link.short_code, &mapping).await
-        }
-    }
-}
-
-/// Result of a redirect operation, containing the response and optional deferred analytics work.
-pub struct RedirectResult {
-    pub response: Response,
-    /// Optional future for analytics logging, to be executed via `ctx.wait_until()`.
-    /// This allows the redirect response to be sent immediately without waiting for D1 writes.
-    pub analytics_future: Option<Pin<Box<dyn Future<Output = ()> + 'static>>>,
-}
-
-/// Handle public short code redirects: GET /{short_code}
-///
-/// Returns a RedirectResult containing the redirect response and an optional
-/// analytics future. The caller should use `Context::wait_until()` to execute
-/// the analytics future after sending the response, avoiding blocking the redirect.
-pub async fn handle_redirect(
-    req: Request,
-    ctx: RouteContext<()>,
-    short_code: String,
-) -> Result<RedirectResult> {
-    let kv = ctx.kv("URL_MAPPINGS")?;
-
-    // Rate limiting: 300 requests per minute per IP
-    let client_ip = get_client_ip(&req);
-    let rate_limit_key = RateLimiter::ip_key("redirect", &client_ip);
-    let rate_limit_config = RateLimitConfig::redirect();
-
-    // Check if KV rate limiting is enabled (default false)
-    let kv_rate_limiting_enabled = ctx
-        .env
-        .var("ENABLE_KV_RATE_LIMITING")
-        .map(|v| v.to_string() == "true")
-        .unwrap_or(false);
-
-    if let Err(err) = RateLimiter::check(
-        &kv,
-        &rate_limit_key,
-        &rate_limit_config,
-        kv_rate_limiting_enabled,
-    )
-    .await
-    {
-        let ip_hash = hash_ip(&client_ip);
-        console_log!(
-            "{}",
-            serde_json::json!({
-                "event": "rate_limit_hit",
-                "endpoint": "redirect",
-                "limit_type": "ip_per_code",
-                "ip_hash": ip_hash,
-                "short_code": short_code,
-                "level": "warn"
-            })
-        );
-        let mut response = Response::error(err.to_error_response(), 429)?;
-        if let Some(retry_after) = err.retry_after() {
-            response
-                .headers_mut()
-                .set("Retry-After", &retry_after.to_string())?;
-        }
-        return Ok(RedirectResult {
-            response,
-            analytics_future: None,
-        });
-    }
-
-    // Look up the link mapping in KV
-    let mapping = kv::get_link_mapping(&kv, &short_code).await?;
-
-    let Some(mapping) = mapping else {
-        let url = Url::parse(&format!("{}/404", get_frontend_url(&ctx.env)))?;
-        return Ok(RedirectResult {
-            response: Response::redirect_with_status(url, 302)?,
-            analytics_future: None,
-        });
-    };
-
-    // Check if link is active
-    if !matches!(mapping.status, LinkStatus::Active) {
-        let url = Url::parse(&format!("{}/404", get_frontend_url(&ctx.env)))?;
-        return Ok(RedirectResult {
-            response: Response::redirect_with_status(url, 302)?,
-            analytics_future: None,
-        });
-    }
-
-    // Check if expired
-    if let Some(expires_at) = mapping.expires_at {
-        let now = now_timestamp();
-
-        if now > expires_at {
-            let url = Url::parse(&format!("{}/404", get_frontend_url(&ctx.env)))?;
-            return Ok(RedirectResult {
-                response: Response::redirect_with_status(url, 302)?,
-                analytics_future: None,
-            });
-        }
-    }
-
-    // Build the redirect URL, applying UTM params and forwarding as needed
-    let mut destination_url = Url::parse(&mapping.destination_url)?;
-
-    // Feature A: apply static UTM params (always, if set on the link)
-    if let Some(ref utm) = mapping.utm_params {
-        let pairs: Vec<(&str, &str)> = [
-            ("utm_source", utm.utm_source.as_deref()),
-            ("utm_medium", utm.utm_medium.as_deref()),
-            ("utm_campaign", utm.utm_campaign.as_deref()),
-            ("utm_term", utm.utm_term.as_deref()),
-            ("utm_content", utm.utm_content.as_deref()),
-            ("utm_ref", utm.utm_ref.as_deref()),
-        ]
-        .into_iter()
-        .filter_map(|(k, v)| v.filter(|s| !s.is_empty()).map(|s| (k, s)))
-        .collect();
-
-        if !pairs.is_empty() {
-            let mut q = destination_url.query_pairs_mut();
-            for (k, v) in pairs {
-                q.append_pair(k, v);
-            }
-        }
-    }
-
-    // Feature B: forward incoming visitor query params (visitor wins on conflict)
-    if mapping.forward_query_params
-        && let Ok(incoming_url) = req.url()
-    {
-        let visitor_pairs: Vec<(String, String)> = incoming_url
-            .query_pairs()
-            .map(|(k, v)| (k.into_owned(), v.into_owned()))
-            .collect();
-        if !visitor_pairs.is_empty() {
-            let mut q = destination_url.query_pairs_mut();
-            for (k, v) in &visitor_pairs {
-                q.append_pair(k, v);
-            }
-        }
-    }
-
-    // Parse redirect type and create appropriate response
-    let redirect_status = mapping.redirect_type.parse::<u16>().unwrap_or(301);
-    let response = Response::redirect_with_status(destination_url, redirect_status)?;
-
-    // Collect analytics data from the request before it's consumed
-    let referrer = req.headers().get("Referer").ok().flatten();
-    let user_agent = req.headers().get("User-Agent").ok().flatten();
-    let country = req.headers().get("CF-IPCountry").ok().flatten();
-    let city = req.headers().get("CF-IPCity").ok().flatten();
-
-    // Prepare deferred analytics work (executed via wait_until after response is sent)
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-    let link_id = mapping.link_id.clone();
-    let now = now_timestamp();
-
-    let analytics_future: Pin<Box<dyn Future<Output = ()> + 'static>> = Box::pin(async move {
-        // Get the full link to extract org_id and check status (no auth check for public redirects)
-        let link = match db::get_link_by_id_no_auth(&db, &link_id).await {
-            Ok(Some(link)) => link,
-            Ok(None) => {
-                console_log!(
-                    "{}",
-                    serde_json::json!({
-                        "event": "analytics_link_not_found",
-                        "link_id": link_id,
-                        "level": "warn"
-                    })
-                );
-                return;
-            }
-            Err(_) => {
-                return;
-            }
-        };
-
-        if !matches!(link.status, LinkStatus::Active) {
-            return;
-        }
-
-        let event = crate::models::AnalyticsEvent {
-            id: None,
-            link_id: link_id.clone(),
-            org_id: link.org_id,
-            timestamp: now,
-            referrer,
-            user_agent,
-            country,
-            city,
-        };
-
-        if let Err(e) = db::log_analytics_event(&db, &event).await {
-            console_log!(
-                "{}",
-                serde_json::json!({
-                    "event": "analytics_event_failed",
-                    "link_id": link_id,
-                    "error": e.to_string(),
-                    "level": "error"
-                })
-            );
-        }
-        if let Err(e) = db::increment_click_count(&db, &link_id).await {
-            console_log!(
-                "{}",
-                serde_json::json!({
-                    "event": "click_count_failed",
-                    "link_id": link_id,
-                    "error": e.to_string(),
-                    "level": "error"
-                })
-            );
-        }
-    });
-
-    Ok(RedirectResult {
-        response,
-        analytics_future: Some(analytics_future),
-    })
-}
-
-#[utoipa::path(
-    post,
-    path = "/api/links",
-    tag = "Links",
-    summary = "Create a link",
-    description = "Creates a new short link for the authenticated organization. Respects monthly tier limits. Optionally accepts a custom short code (Pro+), UTM parameters (Pro+), tags, expiry, and redirect type",
-    request_body(content = CreateLinkRequest, description = "Link creation payload"),
-    responses(
-        (status = 201, description = "Link created", body = Link),
-        (status = 400, description = "Invalid request body or URL"),
-        (status = 401, description = "Unauthorized"),
-        (status = 403, description = "Monthly link limit reached for current tier"),
-        (status = 409, description = "Short code already in use"),
-    ),
-    security(
-        ("Bearer" = []),
-        ("session_cookie" = [])
-    )
-)]
-pub async fn handle_create_link(mut req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    // Authenticate request
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-    let user_id = &user_ctx.user_id;
-    let org_id = &user_ctx.org_id;
-
-    // Rate limiting: 100 link creations per hour per user
-    let kv = ctx.kv("URL_MAPPINGS")?;
-    let rate_limit_key = RateLimiter::user_key("create_link", user_id);
-    let rate_limit_config = RateLimitConfig::link_creation();
-
-    // Check if KV rate limiting is enabled (default false)
-    let kv_rate_limiting_enabled = ctx
-        .env
-        .var("ENABLE_KV_RATE_LIMITING")
-        .map(|v| v.to_string() == "true")
-        .unwrap_or(false);
-
-    if let Err(err) = RateLimiter::check(
-        &kv,
-        &rate_limit_key,
-        &rate_limit_config,
-        kv_rate_limiting_enabled,
-    )
-    .await
-    {
-        console_log!(
-            "{}",
-            serde_json::json!({
-                "event": "rate_limit_hit",
-                "endpoint": "create_link",
-                "limit_type": "user",
-                "user_id": user_id,
-                "level": "warn"
-            })
-        );
-        let mut response = Response::error(err.to_error_response(), 429)?;
-        if let Some(retry_after) = err.retry_after() {
-            response
-                .headers_mut()
-                .set("Retry-After", &retry_after.to_string())?;
-        }
-        response.headers_mut().set(
-            "X-RateLimit-Limit",
-            &rate_limit_config.max_requests.to_string(),
-        )?;
-        return Ok(response);
-    }
-
-    // Tier-based link creation limit (enforced at billing account level)
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-
-    // Get billing account for this org
-    let billing_account = db::get_billing_account_for_org(&db, org_id)
-        .await?
-        .ok_or_else(|| Error::RustError("No billing account found for organization".to_string()))?;
-
-    // Get tier limits from billing account (not org)
-    let tier = Tier::from_str_value(&billing_account.tier);
-    let limits = tier.as_ref().map(|t| t.limits());
-
-    // Check link creation limit for this tier
-    // Quota is shared across all orgs in the billing account
-    if let Some(ref tier_limits) = limits
-        && let Some(max_links) = tier_limits.max_links_per_month
-    {
-        let now = chrono::Utc::now();
-        let year_month = format!("{}-{:02}", now.year(), now.month());
-
-        // Try to increment the counter at billing account level
-        let can_create = db::increment_monthly_counter_for_billing_account(
-            &db,
-            &billing_account.id,
-            &year_month,
-            max_links,
-        )
-        .await?;
-
-        if !can_create {
-            let current_count =
-                db::get_monthly_counter_for_billing_account(&db, &billing_account.id, &year_month)
-                    .await?;
-            let remaining = max_links.saturating_sub(current_count);
-            let message = if remaining > 0 {
-                format!(
-                    "You can create {} more short links this month across all organizations.",
-                    remaining
-                )
-            } else {
-                "You have reached your monthly link limit across all organizations. Upgrade your plan to create more links."
-                    .to_string()
-            };
-            return Response::error(message, 403);
-        }
-    }
-
-    // Parse request body with proper error handling
-    let raw_body: serde_json::Value = match req.json().await {
-        Ok(body) => body,
-        Err(e) => {
-            return Response::error(format!("Invalid JSON: {}", e), 400);
-        }
-    };
-
-    // Validate that only expected fields are present
-    let expected_fields = [
-        "destination_url",
-        "short_code",
-        "title",
-        "expires_at",
-        "tags",
-        "utm_params",
-        "forward_query_params",
-        "redirect_type",
-    ];
-    if let Some(obj) = raw_body.as_object() {
-        for field_name in obj.keys() {
-            if !expected_fields.contains(&field_name.as_str()) {
-                return Response::error(
-                    format!(
-                        "Unknown field '{}'. Expected fields: destination_url, short_code (optional), title (optional), expires_at (optional), tags (optional), utm_params (optional, Pro+), forward_query_params (optional, Pro+), redirect_type (optional, defaults to 301)",
-                        field_name
-                    ),
-                    400,
-                );
-            }
-        }
-    } else {
-        return Response::error("Request body must be a JSON object", 400);
-    }
-
-    // Convert to typed struct
-    let body: CreateLinkRequest = match serde_json::from_value(raw_body) {
-        Ok(body) => body,
-        Err(e) => {
-            return Response::error(format!("Invalid request format: {}", e), 400);
-        }
-    };
-
-    // Validate destination URL
-    let destination_url = match validate_url(&body.destination_url) {
-        Ok(url) => url,
-        Err(e) => {
-            return Response::error(format!("Invalid destination URL: {}", e), 400);
-        }
-    };
-
-    // Check if destination is blacklisted
-    if db::is_destination_blacklisted(&db, &destination_url).await? {
-        return Response::error("Destination URL is blocked", 403);
-    }
-
-    // Validate title length (max 200 characters)
-    if let Some(ref title) = body.title
-        && title.len() > 200
-    {
-        return Response::error("Title must be 200 characters or less", 400);
-    }
-
-    // Check if custom short code is allowed for this tier
-    let allow_custom = limits
-        .as_ref()
-        .map(|l| l.allow_custom_short_code)
-        .unwrap_or(false);
-    if body.short_code.is_some() && !allow_custom {
-        return Response::error(
-            "Custom short codes are not available on the free tier. Upgrade to Pro.",
-            403,
-        );
-    }
-
-    // Check if Pro features (UTM params / query forwarding / redirect type) are allowed
-    let is_pro_or_above = matches!(
-        tier.as_ref(),
-        Some(Tier::Pro) | Some(Tier::Business) | Some(Tier::Unlimited)
-    );
-
-    let wants_pro_features = body
-        .utm_params
-        .as_ref()
-        .map(|u| !u.is_empty())
-        .unwrap_or(false)
-        || body.forward_query_params.is_some()
-        || body.redirect_type != "301"; // 307 is a Pro feature
-    if wants_pro_features && !is_pro_or_above {
-        let error_msg = if body.redirect_type != "301" {
-            "Custom redirect types (307) require a Pro plan or above."
-        } else {
-            "UTM parameters and query parameter forwarding require a Pro plan or above."
-        };
-        return Response::error(error_msg, 403);
-    }
-
-    // Generate or validate short code
-    let short_code = if let Some(custom_code) = body.short_code {
-        match validate_short_code(&custom_code) {
-            Ok(code) => code,
-            Err(e) => {
-                return Response::error(format!("Invalid short code: {}", e), 400);
-            }
-        };
-
-        // Check if already exists
-        let kv = ctx.kv("URL_MAPPINGS")?;
-        if kv::links::short_code_exists(&kv, &custom_code).await? {
-            return Response::error("Short code already in use", 409);
-        }
-
-        custom_code
-    } else {
-        // Generate random code and check for collisions (very rare)
-        let kv = ctx.kv("URL_MAPPINGS")?;
-        let mut code = generate_short_code();
-        let mut attempts = 0;
-
-        while kv::links::short_code_exists(&kv, &code).await? {
-            code = generate_short_code();
-            attempts += 1;
-            if attempts > 10 {
-                return Response::error("Failed to generate unique short code", 500);
-            }
-        }
-
-        code
-    };
-
-    // Validate and normalize tags if provided
-    let normalized_tags = if let Some(tags) = body.tags {
-        match validate_and_normalize_tags(&tags) {
-            Ok(t) => t,
-            Err(e) => return Response::error(e.to_string(), 400),
-        }
-    } else {
-        Vec::new()
-    };
-
-    // Check tag limit for this tier (enforced at billing account level)
-    if let Some(ref tier_limits) = limits
-        && let Some(max_tags) = tier_limits.max_tags
-    {
-        // Get current count of distinct tags in the billing account
-        let current_tag_count = TagRepository::new()
-            .count_distinct_tags_for_billing_account(&db, &billing_account.id)
-            .await?;
-
-        // Find which tags are new to the BA (not already in use)
-        let mut new_tag_count = 0;
-        if !normalized_tags.is_empty() {
-            // Get all existing tags in the BA
-            let existing_tags_query = db.prepare(
-                "SELECT DISTINCT tag_name
-                 FROM link_tags lt
-                 JOIN organizations o ON lt.org_id = o.id
-                 WHERE o.billing_account_id = ?1",
-            );
-            let existing_tags_result = existing_tags_query
-                .bind(&[billing_account.id.clone().into()])?
-                .all()
-                .await?;
-            let existing_tags_set: std::collections::HashSet<String> = existing_tags_result
-                .results::<serde_json::Value>()?
-                .iter()
-                .filter_map(|row| row["tag_name"].as_str().map(|s| s.to_string()))
-                .collect();
-
-            // Count how many of the submitted tags are new
-            new_tag_count = normalized_tags
-                .iter()
-                .filter(|tag| !existing_tags_set.contains(*tag))
-                .count() as i64;
-        }
-
-        // Check if adding these new tags would exceed the limit
-        if current_tag_count + new_tag_count > max_tags {
-            let remaining = max_tags.saturating_sub(current_tag_count);
-            let message = if remaining > 0 {
-                format!(
-                    "You can create {} more tag{} across all organizations. Upgrade your plan to add more tags.",
-                    remaining,
-                    if remaining == 1 { "" } else { "s" }
-                )
-            } else {
-                "You have reached your tag limit across all organizations. Upgrade your plan to create more tags."
-                    .to_string()
-            };
-            return Response::error(message, 403);
-        }
-    }
-
-    // Create link record
-    let link_id = uuid::Uuid::new_v4().to_string();
-    let now = now_timestamp();
-
-    // Normalise UTM params: treat all-empty as None
-    let utm_params = body.utm_params.filter(|u| !u.is_empty());
-
-    let link = Link {
-        id: link_id.clone(),
-        org_id: org_id.to_string(),
-        short_code: short_code.clone(),
-        destination_url: destination_url.clone(),
-        title: body.title,
-        created_by: user_id.to_string(),
-        created_at: now,
-        updated_at: None,
-        expires_at: body.expires_at,
-        status: LinkStatus::Active,
-        click_count: 0,
-        tags: normalized_tags.clone(),
-        utm_params,
-        forward_query_params: body.forward_query_params,
-        redirect_type: body.redirect_type.clone(),
-    };
-
-    // Store in D1
-    db::create_link(&db, &link).await?;
-
-    // Store tags
-    if !normalized_tags.is_empty() {
-        db::set_tags_for_link(&db, &link_id, org_id, &normalized_tags).await?;
-    }
-
-    // Resolve forwarding flag: per-link setting overrides org default
-    let resolved_forward = link.forward_query_params.unwrap_or(false); // org default fetched below if None
-    let resolved_forward = if link.forward_query_params.is_none() {
-        db::get_org_forward_query_params(&db, org_id)
-            .await
-            .unwrap_or(false)
-    } else {
-        resolved_forward
-    };
-
-    // Store in KV
-    let kv = ctx.kv("URL_MAPPINGS")?;
-    let mapping = link.to_mapping(resolved_forward);
-    kv::store_link_mapping(&kv, org_id, &short_code, &mapping).await?;
-
-    Response::from_json(&link)
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/links",
-    tag = "Links",
-    summary = "List links",
-    description = "Returns a paginated list of links for the authenticated organization",
-    params(
-        ("page" = Option<i64>, Query, description = "Page number (default: 1)"),
-        ("limit" = Option<i64>, Query, description = "Items per page (default: 50, max: 100)"),
-        ("tag" = Option<String>, Query, description = "Filter by tag"),
-        ("search" = Option<String>, Query, description = "Search by title or URL"),
-        ("sort" = Option<String>, Query, description = "Sort field: created_at, click_count"),
-        ("order" = Option<String>, Query, description = "Sort order: asc, desc"),
-    ),
-    responses(
-        (status = 200, description = "Paginated list of links"),
-        (status = 401, description = "Unauthorized"),
-    ),
-    security(
-        ("Bearer" = []),
-        ("session_cookie" = [])
-    )
-)]
-pub async fn handle_list_links(req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    // Authenticate request
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-    let org_id = &user_ctx.org_id;
-
-    // Parse pagination params
-    let url = req.url()?;
-    let query = url.query().unwrap_or("");
-
-    let page: i64 = query
-        .split('&')
-        .find(|s| s.starts_with("page="))
-        .and_then(|s| s.split('=').nth(1))
-        .and_then(|s| s.parse().ok())
-        .unwrap_or(1)
-        .max(1); // Ensure page is at least 1
-
-    let limit: i64 = query
-        .split('&')
-        .find(|s| s.starts_with("limit="))
-        .and_then(|s| s.split('=').nth(1))
-        .and_then(|s| s.parse().ok())
-        .unwrap_or(20)
-        .min(100); // Cap at 100 items per page to prevent DoS via unbounded queries
-
-    // Parse search parameter
-    let search = query
-        .split('&')
-        .find(|s| s.starts_with("search="))
-        .and_then(|s| s.split('=').nth(1))
-        .map(|s| urlencoding::decode(s).unwrap_or_default().into_owned())
-        .filter(|s| !s.trim().is_empty() && s.len() <= 100);
-
-    // Parse status filter parameter
-    let status_filter = query
-        .split('&')
-        .find(|s| s.starts_with("status="))
-        .and_then(|s| s.split('=').nth(1))
-        .and_then(|s| match s {
-            "active" | "disabled" => Some(s),
-            _ => None,
-        });
-
-    // Parse sort parameter
-    let sort = query
-        .split('&')
-        .find(|s| s.starts_with("sort="))
-        .and_then(|s| s.split('=').nth(1))
-        .map(|s| match s {
-            "clicks" | "updated" | "title" | "code" => s,
-            _ => "created", // Default: created
-        })
-        .unwrap_or("created");
-
-    // Parse tags filter parameter: ?tags=foo,bar (OR semantics)
-    let tags_filter: Vec<String> = query
-        .split('&')
-        .find(|s| s.starts_with("tags="))
-        .and_then(|s| s.split('=').nth(1))
-        .map(|s| {
-            // Replace + with space before decoding (urlencoding crate doesn't handle + correctly)
-            let s_plus_fixed = s.replace('+', " ");
-            urlencoding::decode(&s_plus_fixed)
-                .unwrap_or_default()
-                .into_owned()
-        })
-        .map(|s| {
-            s.split(',')
-                .map(|t| t.trim().to_string())
-                .filter(|t| !t.is_empty())
-                .collect()
-        })
-        .unwrap_or_default();
-
-    let tags_filter_opt: Option<&[String]> = if tags_filter.is_empty() {
-        None
-    } else {
-        Some(&tags_filter)
-    };
-
-    let offset = (page - 1) * limit;
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-
-    // Get total count, links, and stats using filtered queries
-    // D1 doesn't support parallel queries, so we run them sequentially
-    let total = db::get_links_count_by_org_filtered(
-        &db,
-        org_id,
-        search.as_deref(),
-        status_filter,
-        tags_filter_opt,
-    )
-    .await?;
-
-    let mut links = db::get_links_by_org_filtered(
-        &db,
-        org_id,
-        search.as_deref(),
-        status_filter,
-        sort,
-        limit,
-        offset,
-        tags_filter_opt,
-    )
-    .await?;
-    let stats = db::get_dashboard_stats(&db, org_id).await?;
-
-    // Batch-fetch tags for all returned links
-    let link_ids: Vec<String> = links.iter().map(|l| l.id.clone()).collect();
-    let tags_map = db::get_tags_for_links(&db, &link_ids).await?;
-    for link in &mut links {
-        link.tags = tags_map.get(&link.id).cloned().unwrap_or_default();
-    }
-
-    // Build pagination metadata
-    let pagination = PaginationMeta::new(page, limit, total);
-
-    // Build paginated response with stats
-    let stats_json = serde_json::to_value(&stats)
-        .map_err(|e| Error::RustError(format!("Failed to serialize stats: {}", e)))?;
-    let response = PaginatedResponse::with_stats(links, pagination, stats_json);
-
-    Response::from_json(&response)
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/links/{id}",
-    tag = "Links",
-    summary = "Get a link",
-    description = "Returns details for a single link belonging to the authenticated organization",
-    params(
-        ("id" = String, Path, description = "Link ID"),
-    ),
-    responses(
-        (status = 200, description = "Link details", body = Link),
-        (status = 401, description = "Unauthorized"),
-        (status = 404, description = "Link not found"),
-    ),
-    security(
-        ("Bearer" = []),
-        ("session_cookie" = [])
-    )
-)]
-pub async fn handle_get_link(req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    // Authenticate request
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-    let org_id = &user_ctx.org_id;
-
-    let link_id = ctx
-        .param("id")
-        .ok_or_else(|| Error::RustError("Missing link ID".to_string()))?;
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-    let link = db::get_link_by_id(&db, link_id, org_id).await?;
-
-    match link {
-        Some(mut link) => {
-            link.tags = db::get_tags_for_link(&db, &link.id).await?;
-            Response::from_json(&link)
-        }
-        None => Response::error("Link not found", 404),
-    }
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/links/by-code/{code}",
-    tag = "Links",
-    summary = "Get a link by short code",
-    description = "Looks up a link by its short code instead of its internal ID",
-    params(
-        ("code" = String, Path, description = "Short code of the link"),
-    ),
-    responses(
-        (status = 200, description = "Link details", body = Link),
-        (status = 401, description = "Unauthorized"),
-        (status = 404, description = "Link not found"),
-    ),
-    security(
-        ("Bearer" = []),
-        ("session_cookie" = [])
-    )
-)]
-pub async fn handle_get_link_by_code(req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    // Authenticate request
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-    let org_id = &user_ctx.org_id;
-
-    let short_code = ctx
-        .param("code")
-        .ok_or_else(|| Error::RustError("Missing short code".to_string()))?;
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-    let link = db::get_link_by_short_code(&db, short_code, org_id).await?;
-
-    match link {
-        Some(link) => Response::from_json(&link),
-        None => Response::error("Link not found", 404),
-    }
-}
-
-#[utoipa::path(
-    delete,
-    path = "/api/links/{id}",
-    tag = "Links",
-    summary = "Delete a link",
-    description = "Soft-deletes a link belonging to the authenticated organization and removes its KV mapping so it stops redirecting",
-    params(
-        ("id" = String, Path, description = "Link ID"),
-    ),
-    responses(
-        (status = 200, description = "Link deleted"),
-        (status = 401, description = "Unauthorized"),
-        (status = 404, description = "Link not found"),
-    ),
-    security(
-        ("Bearer" = []),
-        ("session_cookie" = [])
-    )
-)]
-pub async fn handle_delete_link(req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    // Authenticate request
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-    let org_id = &user_ctx.org_id;
-
-    let link_id = ctx
-        .param("id")
-        .ok_or_else(|| Error::RustError("Missing link ID".to_string()))?;
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-
-    // Get link first to get short_code
-    let link = db::get_link_by_id(&db, link_id, org_id).await?;
-
-    let Some(link) = link else {
-        return Response::error("Link not found", 404);
-    };
-
-    // Delete tags first (FK constraint)
-    db::delete_tags_for_link(&db, link_id).await?;
-
-    // Hard delete from D1 (frees up short code)
-    db::hard_delete_link(&db, link_id, org_id).await?;
-
-    // Hard delete from KV
-    let kv = ctx.kv("URL_MAPPINGS")?;
-    kv::delete_link_mapping(&kv, org_id, &link.short_code).await?;
-
-    Response::empty()
-}
-
-#[utoipa::path(
-    put,
-    path = "/api/links/{id}",
-    tag = "Links",
-    summary = "Update a link",
-    description = "Updates a link's destination URL, title, tags, expiry, UTM parameters, redirect type, or forward-query-params setting. Updates are written to both D1 and KV atomically",
-    params(
-        ("id" = String, Path, description = "Link ID"),
-    ),
-    request_body(content = UpdateLinkRequest, description = "Fields to update (all optional)"),
-    responses(
-        (status = 200, description = "Updated link", body = Link),
-        (status = 400, description = "Invalid request body"),
-        (status = 401, description = "Unauthorized"),
-        (status = 404, description = "Link not found"),
-    ),
-    security(
-        ("Bearer" = []),
-        ("session_cookie" = [])
-    )
-)]
-pub async fn handle_update_link(mut req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    // Authenticate
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-
-    // Extract link ID from route
-    let link_id = match ctx.param("id") {
-        Some(id) => id.to_string(),
-        None => return Response::error("Missing link ID", 400),
-    };
-
-    // Parse request body
-    let update_req: UpdateLinkRequest = match req.json().await {
-        Ok(req) => req,
-        Err(_) => return Response::error("Invalid request body", 400),
-    };
-
-    // Validate destination URL if provided
-    if let Some(url) = &update_req.destination_url {
-        if let Err(e) = validate_url(url) {
-            return Response::error(format!("Invalid URL: {}", e), 400);
-        }
-
-        // Check if destination is blacklisted
-        let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-        if db::is_destination_blacklisted(&db, url).await? {
-            return Response::error("Destination URL is blocked", 403);
-        }
-    }
-
-    // Validate title length if provided (max 200 characters)
-    if let Some(ref title) = update_req.title
-        && title.len() > 200
-    {
-        return Response::error("Title must be 200 characters or less", 400);
-    }
-
-    // Validate expiration date if provided
-    if let Some(expires_at) = update_req.expires_at {
-        let now = now_timestamp();
-        if expires_at <= now {
-            return Response::error("Expiration date must be in the future", 400);
-        }
-    }
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-    let kv = ctx.kv("URL_MAPPINGS")?;
-
-    // Get existing link to verify ownership
-    let _existing_link = match db::get_link_by_id(&db, &link_id, &user_ctx.org_id).await? {
-        Some(link) => link,
-        None => return Response::error("Link not found", 404),
-    };
-
-    // Tier check for Pro-only features in update
-    let billing_account_update = db::get_billing_account_for_org(&db, &user_ctx.org_id)
-        .await?
-        .ok_or_else(|| Error::RustError("No billing account found for organization".to_string()))?;
-    let tier_update = Tier::from_str_value(&billing_account_update.tier);
-    let is_pro_or_above_update = matches!(
-        tier_update.as_ref(),
-        Some(Tier::Pro) | Some(Tier::Business) | Some(Tier::Unlimited)
-    );
-
-    let wants_pro_features_update = update_req
-        .utm_params
-        .as_ref()
-        .map(|u| !u.is_empty())
-        .unwrap_or(false)
-        || update_req.forward_query_params.is_some()
-        || (update_req.redirect_type.is_some()
-            && update_req.redirect_type.as_deref() != Some("301"));
-    if wants_pro_features_update && !is_pro_or_above_update {
-        let error_msg = if update_req.redirect_type.is_some()
-            && update_req.redirect_type.as_deref() != Some("301")
-        {
-            "Custom redirect types (307) require a Pro plan or above."
-        } else {
-            "UTM parameters and query parameter forwarding require a Pro plan or above."
-        };
-        return Response::error(error_msg, 403);
-    }
-
-    // Normalise UTM: None means "not provided" (no change); Some(empty) means clear it
-    let utm_json_for_db: Option<Option<String>> = update_req.utm_params.as_ref().map(|u| {
-        if u.is_empty() {
-            None
-        } else {
-            u.to_json_string()
-        }
-    });
-
-    // Update in D1
-    let mut updated_link = db::update_link(
-        &db,
-        &link_id,
-        &user_ctx.org_id,
-        update_req.destination_url.as_deref(),
-        update_req.title.as_deref(),
-        update_req.status.as_ref().map(|s| s.as_str()),
-        update_req.expires_at,
-        utm_json_for_db.as_ref().map(|o| o.as_deref()),
-        update_req.forward_query_params.map(Some),
-        update_req.redirect_type.as_deref(),
-    )
-    .await?;
-
-    // Update tags if provided
-    if let Some(tags) = update_req.tags {
-        let normalized_tags = match validate_and_normalize_tags(&tags) {
-            Ok(t) => t,
-            Err(e) => return Response::error(e.to_string(), 400),
-        };
-
-        // Check tag limit for this tier (enforced at billing account level)
-        let tier_limits = tier_update.as_ref().map(|t| t.limits());
-        if let Some(ref limits) = tier_limits
-            && let Some(max_tags) = limits.max_tags
-        {
-            // Get current count of distinct tags in the billing account
-            let current_tag_count = TagRepository::new()
-                .count_distinct_tags_for_billing_account(&db, &billing_account_update.id)
-                .await?;
-
-            // Get existing tags for this link to see which ones are being removed
-            let existing_link_tags = db::get_tags_for_link(&db, &link_id).await?;
-            let existing_tags_set: std::collections::HashSet<String> =
-                existing_link_tags.into_iter().collect();
-            let new_tags_set: std::collections::HashSet<String> =
-                normalized_tags.iter().cloned().collect();
-
-            // Get all existing tags in the BA
-            let existing_ba_tags_query = db.prepare(
-                "SELECT DISTINCT tag_name
-                 FROM link_tags lt
-                 JOIN organizations o ON lt.org_id = o.id
-                 WHERE o.billing_account_id = ?1",
-            );
-            let existing_ba_tags_result = existing_ba_tags_query
-                .bind(&[billing_account_update.id.clone().into()])?
-                .all()
-                .await?;
-            let existing_ba_tags_set: std::collections::HashSet<String> = existing_ba_tags_result
-                .results::<serde_json::Value>()?
-                .iter()
-                .filter_map(|row| row["tag_name"].as_str().map(|s| s.to_string()))
-                .collect();
-
-            // Calculate net change in distinct tags
-            // Tags being removed from this link that aren't used elsewhere in the BA
-            let tags_being_removed: std::collections::HashSet<String> = existing_tags_set
-                .difference(&new_tags_set)
-                .cloned()
-                .collect();
-
-            // Tags being added that don't exist elsewhere in the BA
-            let tags_being_added: std::collections::HashSet<String> = new_tags_set
-                .difference(&existing_tags_set)
-                .cloned()
-                .collect();
-
-            // Count how many removed tags are actually disappearing from the BA
-            let mut disappearing_count = 0;
-            for tag in &tags_being_removed {
-                // Check if this tag is used anywhere else in the BA (excluding this link)
-                let usage_query = db.prepare(
-                    "SELECT COUNT(*) as count
-                     FROM link_tags lt
-                     JOIN organizations o ON lt.org_id = o.id
-                     WHERE o.billing_account_id = ?1 AND lt.tag_name = ?2 AND lt.link_id != ?3",
-                );
-                let usage_result = usage_query
-                    .bind(&[
-                        billing_account_update.id.clone().into(),
-                        tag.as_str().into(),
-                        link_id.as_str().into(),
-                    ])?
-                    .first::<serde_json::Value>(None)
-                    .await?;
-                let usage_count = if let Some(result) = usage_result {
-                    result["count"].as_f64().unwrap_or(0.0) as i64
-                } else {
-                    0
-                };
-
-                if usage_count == 0 {
-                    disappearing_count += 1;
-                }
-            }
-
-            // Count how many added tags are truly new to the BA
-            let new_to_ba_count = tags_being_added
-                .iter()
-                .filter(|tag| !existing_ba_tags_set.contains(*tag))
-                .count() as i64;
-
-            // Net change in tag count
-            let net_change = new_to_ba_count - disappearing_count;
-
-            // Check if this would exceed the limit
-            if current_tag_count + net_change > max_tags {
-                let remaining = max_tags.saturating_sub(current_tag_count);
-                let message = if remaining > 0 {
-                    format!(
-                        "You can create {} more tag{} across all organizations. Upgrade your plan to add more tags.",
-                        remaining,
-                        if remaining == 1 { "" } else { "s" }
-                    )
-                } else {
-                    "You have reached your tag limit across all organizations. Upgrade your plan to create more tags."
-                        .to_string()
-                };
-                return Response::error(message, 403);
-            }
-        }
-
-        db::set_tags_for_link(&db, &link_id, &user_ctx.org_id, &normalized_tags).await?;
-        updated_link.tags = normalized_tags;
-    } else {
-        updated_link.tags = db::get_tags_for_link(&db, &link_id).await?;
-    }
-
-    // Update KV if any redirect-affecting field changed
-    let kv_needs_update = update_req.destination_url.is_some()
-        || update_req.status.is_some()
-        || update_req.utm_params.is_some()
-        || update_req.forward_query_params.is_some();
-    if kv_needs_update {
-        sync_link_mapping_from_link(&db, &kv, &updated_link).await?;
-    }
-
-    Response::from_json(&updated_link)
-}
-
-// ─── CSV helpers ─────────────────────────────────────────────────────────────
-
-/// Escape a single CSV field: wraps in double-quotes if the value contains
-/// a comma, double-quote, or newline; doubles any embedded double-quotes.
-fn csv_escape(value: &str) -> String {
-    if value.contains(',') || value.contains('"') || value.contains('\n') || value.contains('\r') {
-        format!("\"{}\"", value.replace('"', "\"\""))
-    } else {
-        value.to_string()
-    }
-}
-
-// ─── Import / Export structs ─────────────────────────────────────────────────
-
-#[derive(Debug, serde::Deserialize)]
-struct ImportLinkRow {
-    destination_url: String,
-    short_code: Option<String>,
-    title: Option<String>,
-    tags: Option<Vec<String>>,
-    expires_at: Option<i64>,
-}
-
-#[derive(Debug, serde::Deserialize)]
-struct ImportRequest {
-    links: Vec<ImportLinkRow>,
-}
-
-#[derive(Debug, serde::Serialize)]
-struct ImportError {
-    row: usize,
-    destination_url: String,
-    reason: String,
-}
-
-#[derive(Debug, serde::Serialize)]
-struct ImportWarning {
-    row: usize,
-    destination_url: String,
-    reason: String,
-}
-
-#[derive(Debug, serde::Serialize)]
-struct ImportResponse {
-    created: usize,
-    skipped: usize,
-    failed: usize,
-    errors: Vec<ImportError>,
-    warnings: Vec<ImportWarning>,
-}
-
-// ─── Export handler ───────────────────────────────────────────────────────────
-
-#[utoipa::path(
-    get,
-    path = "/api/links/export",
-    tag = "Links",
-    summary = "Export links as CSV",
-    description = "Exports all active links for the authenticated organization as a CSV file. Returns a text/csv response with a Content-Disposition attachment header",
-    responses(
-        (status = 200, description = "CSV file download"),
-        (status = 401, description = "Unauthorized"),
-    ),
-    security(
-        ("Bearer" = []),
-        ("session_cookie" = [])
-    )
-)]
-pub async fn handle_export_links(req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-    let org_id = &user_ctx.org_id;
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-
-    let mut links = db::get_all_links_for_org_export(&db, org_id).await?;
-
-    let link_ids: Vec<String> = links.iter().map(|l| l.id.clone()).collect();
-    let tags_map = db::get_tags_for_links(&db, &link_ids).await?;
-    for link in &mut links {
-        link.tags = tags_map.get(&link.id).cloned().unwrap_or_default();
-    }
-
-    let mut csv = String::from(
-        "short_code,destination_url,title,tags,status,click_count,created_at,expires_at,utm_source,utm_medium,utm_campaign,utm_term,utm_content,utm_ref,forward_query_params\n",
-    );
-
-    for link in &links {
-        let title = link.title.as_deref().unwrap_or("");
-        let tags_str = link.tags.join("|");
-        let created_at = chrono::DateTime::from_timestamp(link.created_at, 0)
-            .map(|dt: chrono::DateTime<chrono::Utc>| dt.format("%Y-%m-%dT%H:%M:%SZ").to_string())
-            .unwrap_or_default();
-        let expires_at = link
-            .expires_at
-            .and_then(|ts| chrono::DateTime::from_timestamp(ts, 0))
-            .map(|dt: chrono::DateTime<chrono::Utc>| dt.format("%Y-%m-%dT%H:%M:%SZ").to_string())
-            .unwrap_or_default();
-        let utm_source = link
-            .utm_params
-            .as_ref()
-            .and_then(|u| u.utm_source.as_deref())
-            .unwrap_or("");
-        let utm_medium = link
-            .utm_params
-            .as_ref()
-            .and_then(|u| u.utm_medium.as_deref())
-            .unwrap_or("");
-        let utm_campaign = link
-            .utm_params
-            .as_ref()
-            .and_then(|u| u.utm_campaign.as_deref())
-            .unwrap_or("");
-        let utm_term = link
-            .utm_params
-            .as_ref()
-            .and_then(|u| u.utm_term.as_deref())
-            .unwrap_or("");
-        let utm_content = link
-            .utm_params
-            .as_ref()
-            .and_then(|u| u.utm_content.as_deref())
-            .unwrap_or("");
-        let utm_ref = link
-            .utm_params
-            .as_ref()
-            .and_then(|u| u.utm_ref.as_deref())
-            .unwrap_or("");
-        let forward_query = link
-            .forward_query_params
-            .map(|v| if v { "true" } else { "false" })
-            .unwrap_or("");
-
-        csv.push_str(&format!(
-            "{},{},{},{},{},{},{},{},{},{},{},{},{},{},{}\n",
-            csv_escape(&link.short_code),
-            csv_escape(&link.destination_url),
-            csv_escape(title),
-            csv_escape(&tags_str),
-            csv_escape(link.status.as_str()),
-            link.click_count,
-            created_at,
-            expires_at,
-            csv_escape(utm_source),
-            csv_escape(utm_medium),
-            csv_escape(utm_campaign),
-            csv_escape(utm_term),
-            csv_escape(utm_content),
-            csv_escape(utm_ref),
-            forward_query,
-        ));
-    }
-
-    let date_str = chrono::Utc::now().format("%Y-%m-%d").to_string();
-    let filename = format!("rushomon-links-{}.csv", date_str);
-    let mut response = Response::ok(csv)?;
-    response
-        .headers_mut()
-        .set("Content-Type", "text/csv; charset=utf-8")?;
-    response.headers_mut().set(
-        "Content-Disposition",
-        &format!("attachment; filename=\"{}\"", filename),
-    )?;
-    Ok(response)
-}
-
-// ─── Import handler ───────────────────────────────────────────────────────────
-
-#[utoipa::path(
-    post,
-    path = "/api/links/import",
-    tag = "Links",
-    summary = "Import links from CSV",
-    description = "Bulk-imports links from a CSV payload. Accepts a JSON array of rows parsed from CSV. Each row must have at least a destination_url. Returns counts of created, skipped (duplicate short codes), and failed rows",
-    responses(
-        (status = 200, description = "Import result with created/skipped/failed counts"),
-        (status = 400, description = "Invalid request body"),
-        (status = 401, description = "Unauthorized"),
-    ),
-    security(
-        ("Bearer" = []),
-        ("session_cookie" = [])
-    )
-)]
-pub async fn handle_import_links(mut req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-    let user_id = &user_ctx.user_id;
-    let org_id = &user_ctx.org_id;
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-
-    let billing_account = db::get_billing_account_for_org(&db, org_id)
-        .await?
-        .ok_or_else(|| Error::RustError("No billing account found for organization".to_string()))?;
-    let tier = Tier::from_str_value(&billing_account.tier);
-    let limits = tier.as_ref().map(|t| t.limits());
-    let is_pro_or_above = matches!(
-        tier.as_ref(),
-        Some(Tier::Pro) | Some(Tier::Business) | Some(Tier::Unlimited)
-    );
-
-    let body: ImportRequest = match req.json().await {
-        Ok(b) => b,
-        Err(_) => return Response::error("Invalid JSON body", 400),
-    };
-
-    if body.links.is_empty() {
-        return Response::from_json(&ImportResponse {
-            created: 0,
-            skipped: 0,
-            failed: 0,
-            errors: vec![],
-            warnings: vec![],
-        });
-    }
-
-    if body.links.len() > 50 {
-        return Response::error("Maximum 50 links per import batch", 400);
-    }
-
-    let kv = ctx.kv("URL_MAPPINGS")?;
-    let now = now_timestamp();
-    let year_month = {
-        let dt = chrono::Utc::now();
-        format!("{}-{:02}", dt.year(), dt.month())
-    };
-
-    let mut created: usize = 0;
-    let mut skipped: usize = 0;
-    let mut failed: usize = 0;
-    let mut errors: Vec<ImportError> = Vec::new();
-    let mut warnings: Vec<ImportWarning> = Vec::new();
-
-    for (idx, row) in body.links.iter().enumerate() {
-        let row_num = idx + 1;
-
-        // Validate destination URL
-        let destination_url = match validate_url(&row.destination_url) {
-            Ok(url) => url,
-            Err(e) => {
-                failed += 1;
-                errors.push(ImportError {
-                    row: row_num,
-                    destination_url: row.destination_url.clone(),
-                    reason: format!("Invalid URL: {}", e),
-                });
-                continue;
-            }
-        };
-
-        // Check blacklist
-        if db::is_destination_blacklisted(&db, &destination_url).await? {
-            failed += 1;
-            errors.push(ImportError {
-                row: row_num,
-                destination_url: destination_url.clone(),
-                reason: "Destination URL is blocked".to_string(),
-            });
-            continue;
-        }
-
-        // Monthly quota check (increment counter; rolls back nothing on partial failure)
-        if let Some(ref tier_limits) = limits
-            && let Some(max_links) = tier_limits.max_links_per_month
-        {
-            let can_create = db::increment_monthly_counter_for_billing_account(
-                &db,
-                &billing_account.id,
-                &year_month,
-                max_links,
-            )
-            .await?;
-            if !can_create {
-                failed += 1;
-                errors.push(ImportError {
-                    row: row_num,
-                    destination_url: destination_url.clone(),
-                    reason: "Monthly link limit reached".to_string(),
-                });
-                continue;
-            }
-        }
-
-        // Determine short code
-        let short_code: String;
-        if is_pro_or_above && let Some(provided_code) = row.short_code.as_ref() {
-            if let Err(e) = validate_short_code(provided_code) {
-                skipped += 1;
-                errors.push(ImportError {
-                    row: row_num,
-                    destination_url: destination_url.clone(),
-                    reason: format!("Invalid short code: {}", e),
-                });
-                continue;
-            }
-
-            // Try provided code, then auto-suffix on conflict (promo → promo-1 → … → promo-10)
-            let mut resolved: Option<String> = None;
-            for attempt in 0u32..=10 {
-                let candidate = if attempt == 0 {
-                    provided_code.clone()
-                } else {
-                    format!("{}-{}", provided_code, attempt)
-                };
-                if !kv::links::short_code_exists(&kv, &candidate).await? {
-                    resolved = Some(candidate);
-                    break;
-                }
-            }
-
-            match resolved {
-                Some(c) => short_code = c,
-                None => {
-                    // All suffix attempts exhausted — fall back to a random code
-                    let mut fallback: Option<String> = None;
-                    for _ in 0..10u32 {
-                        let candidate = generate_short_code();
-                        if !kv::links::short_code_exists(&kv, &candidate).await? {
-                            fallback = Some(candidate);
-                            break;
-                        }
-                    }
-                    match fallback {
-                        Some(c) => {
-                            warnings.push(ImportWarning {
-                                row: row_num,
-                                destination_url: destination_url.clone(),
-                                reason: format!(
-                                    "Short code '{}' conflicted with an existing link; a random code was assigned",
-                                    provided_code
-                                ),
-                            });
-                            short_code = c;
-                        }
-                        None => {
-                            failed += 1;
-                            errors.push(ImportError {
-                                row: row_num,
-                                destination_url: destination_url.clone(),
-                                reason: "Failed to generate a unique short code after conflict"
-                                    .to_string(),
-                            });
-                            continue;
-                        }
-                    }
-                }
-            }
-        } else {
-            // Free tier OR Pro without provided code: auto-generate
-            let mut resolved: Option<String> = None;
-            for _ in 0..10u32 {
-                let candidate = generate_short_code();
-                if !kv::links::short_code_exists(&kv, &candidate).await? {
-                    resolved = Some(candidate);
-                    break;
-                }
-            }
-            match resolved {
-                Some(c) => short_code = c,
-                None => {
-                    failed += 1;
-                    errors.push(ImportError {
-                        row: row_num,
-                        destination_url: destination_url.clone(),
-                        reason: "Failed to generate unique short code".to_string(),
-                    });
-                    continue;
-                }
-            }
-        }
-
-        // Validate and normalize tags (silently drop bad tags)
-        let mut normalized_tags = if let Some(ref tags) = row.tags {
-            validate_and_normalize_tags(tags).unwrap_or_default()
-        } else {
-            Vec::new()
-        };
-
-        // Check tag limit for this tier (enforced at billing account level)
-        if let Some(ref tier_limits) = limits
-            && let Some(max_tags) = tier_limits.max_tags
-        {
-            // Get current count of distinct tags in the billing account
-            let current_tag_count = TagRepository::new()
-                .count_distinct_tags_for_billing_account(&db, &billing_account.id)
-                .await?;
-
-            // Find which tags are new to the BA (not already in use)
-            let mut new_tag_count = 0;
-            if !normalized_tags.is_empty() {
-                // Get all existing tags in the BA
-                let existing_tags_query = db.prepare(
-                    "SELECT DISTINCT tag_name
-                     FROM link_tags lt
-                     JOIN organizations o ON lt.org_id = o.id
-                     WHERE o.billing_account_id = ?1",
-                );
-                let existing_tags_result = existing_tags_query
-                    .bind(&[billing_account.id.clone().into()])?
-                    .all()
-                    .await?;
-                let existing_tags_set: std::collections::HashSet<String> = existing_tags_result
-                    .results::<serde_json::Value>()?
-                    .iter()
-                    .filter_map(|row| row["tag_name"].as_str().map(|s| s.to_string()))
-                    .collect();
-
-                // Count how many of the submitted tags are new
-                new_tag_count = normalized_tags
-                    .iter()
-                    .filter(|tag| !existing_tags_set.contains(*tag))
-                    .count() as i64;
-            }
-
-            // Check if adding these new tags would exceed the limit
-            if current_tag_count + new_tag_count > max_tags {
-                // For imports, we'll skip the row but allow the import to continue
-                skipped += 1;
-                warnings.push(ImportWarning {
-                    row: row_num,
-                    destination_url: destination_url.clone(),
-                    reason: format!(
-                        "Tags skipped: would exceed tag limit ({} max). Consider upgrading your plan.",
-                        max_tags
-                    ),
-                });
-                // Clear tags to avoid exceeding limit
-                normalized_tags.clear();
-            }
-        }
-
-        // Clamp title to 200 chars
-        let title = row.title.as_ref().and_then(|t| {
-            let trimmed = t.trim().to_string();
-            if trimmed.is_empty() || trimmed.len() > 200 {
-                None
-            } else {
-                Some(trimmed)
-            }
-        });
-
-        let link_id = uuid::Uuid::new_v4().to_string();
-        let link = Link {
-            id: link_id.clone(),
-            org_id: org_id.to_string(),
-            short_code: short_code.clone(),
-            destination_url: destination_url.clone(),
-            title,
-            created_by: user_id.to_string(),
-            created_at: now,
-            updated_at: None,
-            expires_at: row.expires_at,
-            status: crate::models::link::LinkStatus::Active,
-            click_count: 0,
-            tags: normalized_tags.clone(),
-            utm_params: None,
-            forward_query_params: None,
-            redirect_type: "301".to_string(), // Default for imported links
-        };
-
-        db::create_link(&db, &link).await?;
-
-        if !normalized_tags.is_empty() {
-            db::set_tags_for_link(&db, &link_id, org_id, &normalized_tags).await?;
-        }
-
-        let mapping = link.to_mapping(false);
-        kv::store_link_mapping(&kv, org_id, &short_code, &mapping).await?;
-
-        created += 1;
-    }
-
-    Response::from_json(&ImportResponse {
-        created,
-        skipped,
-        failed,
-        errors,
-        warnings,
-    })
 }
 
 /// Shared rate-limit + redirect_uri setup for OAuth initiation handlers
@@ -2075,343 +417,6 @@ pub async fn handle_admin_reset_monthly_counter(
         }
     }
 }
-
-#[utoipa::path(
-    get,
-    path = "/api/admin/links",
-    tag = "Admin",
-    summary = "List all links (moderation)",
-    params(
-        ("page" = Option<i64>, Query, description = "Page number"),
-        ("limit" = Option<i64>, Query, description = "Items per page (max 100)"),
-        ("org" = Option<String>, Query, description = "Filter by org ID"),
-        ("email" = Option<String>, Query, description = "Filter by creator email"),
-        ("domain" = Option<String>, Query, description = "Filter by destination domain"),
-    ),
-    responses(
-        (status = 200, description = "Paginated list of all links"),
-        (status = 401, description = "Unauthorized"),
-        (status = 403, description = "Admin required"),
-    ),
-    security(("Bearer" = []), ("session_cookie" = []))
-)]
-pub async fn handle_admin_list_links(req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-
-    if let Err(e) = auth::require_admin(&user_ctx) {
-        return Ok(e.into_response());
-    }
-
-    // Parse pagination and filter params
-    let url = req.url()?;
-    let page: i64 = url
-        .query()
-        .and_then(|q| {
-            q.split('&')
-                .find(|s| s.starts_with("page="))
-                .and_then(|s| s.split('=').nth(1))
-                .and_then(|s| s.parse().ok())
-        })
-        .unwrap_or(1);
-
-    let limit: i64 = url
-        .query()
-        .and_then(|q| {
-            q.split('&')
-                .find(|s| s.starts_with("limit="))
-                .and_then(|s| s.split('=').nth(1))
-                .and_then(|s| s.parse().ok())
-        })
-        .unwrap_or(50)
-        .min(100);
-
-    let offset = (page - 1) * limit;
-
-    // Parse filters
-    let org_filter = url.query().and_then(|q| {
-        q.split('&')
-            .find(|s| s.starts_with("org="))
-            .and_then(|s| s.split('=').nth(1))
-    });
-
-    let email_filter = url.query().and_then(|q| {
-        q.split('&')
-            .find(|s| s.starts_with("email="))
-            .and_then(|s| s.split('=').nth(1))
-    });
-
-    let domain_filter = url.query().and_then(|q| {
-        q.split('&')
-            .find(|s| s.starts_with("domain="))
-            .and_then(|s| s.split('=').nth(1))
-    });
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-    let kv = ctx.kv("URL_MAPPINGS")?;
-    let links = db::get_all_links_admin(
-        &db,
-        &kv,
-        limit,
-        offset,
-        org_filter,
-        email_filter,
-        domain_filter,
-    )
-    .await?;
-    let total = db::get_all_links_admin_count(&db, org_filter, email_filter, domain_filter).await?;
-
-    Response::from_json(&serde_json::json!({
-        "links": links,
-        "total": total,
-        "page": page,
-        "limit": limit,
-    }))
-}
-
-#[utoipa::path(
-    put,
-    path = "/api/admin/links/{id}",
-    tag = "Admin",
-    summary = "Update link status",
-    params(("id" = String, Path, description = "Link ID")),
-    responses(
-        (status = 200, description = "Status updated"),
-        (status = 400, description = "Invalid status value"),
-        (status = 401, description = "Unauthorized"),
-        (status = 403, description = "Admin required"),
-    ),
-    security(("Bearer" = []), ("session_cookie" = []))
-)]
-pub async fn handle_admin_update_link_status(
-    mut req: Request,
-    ctx: RouteContext<()>,
-) -> Result<Response> {
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-
-    if let Err(e) = auth::require_admin(&user_ctx) {
-        return Ok(e.into_response());
-    }
-
-    let link_id = ctx
-        .param("id")
-        .ok_or_else(|| Error::RustError("Missing link ID".to_string()))?
-        .to_string();
-
-    // Parse request body
-    let body: serde_json::Value = match req.json().await {
-        Ok(body) => body,
-        Err(e) => return Response::error(format!("Invalid JSON: {}", e), 400),
-    };
-
-    let status = match body.get("status").and_then(|s| s.as_str()) {
-        Some(s) if s == "active" || s == "disabled" || s == "blocked" => s.to_string(),
-        Some(_) => {
-            return Response::error(
-                "Invalid status. Must be 'active', 'disabled', or 'blocked'",
-                400,
-            );
-        }
-        None => return Response::error("Missing 'status' field", 400),
-    };
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-
-    // Update link status using the same function as regular endpoint
-    db::update_link_status_by_id(&db, &link_id, &status).await?;
-
-    // Sync KV using the shared function (now that get_link_by_id_no_auth_all is fixed)
-    let kv = ctx.kv("URL_MAPPINGS")?;
-    match db::get_link_by_id_no_auth_all(&db, &link_id).await? {
-        Some(updated_link) => {
-            sync_link_mapping_from_link(&db, &kv, &updated_link).await?;
-        }
-        None => {
-            console_log!(
-                "{}",
-                serde_json::json!({
-                    "event": "admin_link_sync_not_found",
-                    "link_id": link_id,
-                    "level": "critical"
-                })
-            );
-        }
-    }
-
-    // Auto-resolve all pending reports for this link if status is being changed to disabled/blocked
-    if (status == "disabled" || status == "blocked")
-        && let Err(e) = db::resolve_reports_for_link(
-            &db,
-            &link_id,
-            "reviewed",
-            &format!("Action taken: Link {}", status),
-            &user_ctx.user_id,
-        )
-        .await
-    {
-        console_log!(
-            "{}",
-            serde_json::json!({
-                "event": "reports_resolve_failed",
-                "link_id": link_id,
-                "error": e.to_string(),
-                "level": "error"
-            })
-        );
-    }
-
-    Response::from_json(&serde_json::json!({
-        "success": true,
-        "message": format!("Link status updated to {}", status)
-    }))
-}
-
-#[utoipa::path(
-    delete,
-    path = "/api/admin/links/{id}",
-    tag = "Admin",
-    summary = "Hard delete a link",
-    params(("id" = String, Path, description = "Link ID")),
-    responses(
-        (status = 200, description = "Link deleted"),
-        (status = 401, description = "Unauthorized"),
-        (status = 403, description = "Admin required"),
-        (status = 404, description = "Link not found"),
-    ),
-    security(("Bearer" = []), ("session_cookie" = []))
-)]
-pub async fn handle_admin_delete_link(req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-
-    if let Err(e) = auth::require_admin(&user_ctx) {
-        return Ok(e.into_response());
-    }
-
-    let link_id = ctx
-        .param("id")
-        .ok_or_else(|| Error::RustError("Missing link ID".to_string()))?
-        .to_string();
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-
-    // Get link first to get short_code for KV deletion
-    let link = match db::get_link_by_id_no_auth_all(&db, &link_id).await? {
-        Some(link) => link,
-        None => return Response::error("Link not found", 404),
-    };
-
-    // Hard delete from D1
-    db::hard_delete_link(&db, &link_id, &link.org_id).await?;
-
-    // Delete from KV
-    let kv = ctx.kv("URL_MAPPINGS")?;
-    kv::delete_link_mapping(&kv, &link.org_id, &link.short_code).await?;
-
-    Response::from_json(&serde_json::json!({
-        "success": true,
-        "message": "Link deleted successfully"
-    }))
-}
-
-#[utoipa::path(
-    post,
-    path = "/api/admin/links/{id}/sync-kv",
-    tag = "Admin",
-    summary = "Re-sync link KV entry",
-    params(("id" = String, Path, description = "Link ID")),
-    responses(
-        (status = 200, description = "KV entry synced"),
-        (status = 401, description = "Unauthorized"),
-        (status = 403, description = "Admin required"),
-        (status = 404, description = "Link not found"),
-    ),
-    security(("Bearer" = []), ("session_cookie" = []))
-)]
-pub async fn handle_admin_sync_link_kv(req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-
-    if let Err(e) = auth::require_admin(&user_ctx) {
-        return Ok(e.into_response());
-    }
-
-    let link_id = ctx
-        .param("id")
-        .ok_or_else(|| Error::RustError("Missing link ID".to_string()))?
-        .to_string();
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-    let kv = ctx.kv("URL_MAPPINGS")?;
-
-    // Get link from D1
-    let link = match db::get_link_by_id_no_auth_all(&db, &link_id).await? {
-        Some(link) => link,
-        None => return Response::error("Link not found", 404),
-    };
-
-    // Re-sync based on D1 status (treat D1 as source of truth)
-    match link.status.as_str() {
-        "active" => {
-            // Link should be active in KV - recreate/update the KV entry
-            let link_model = crate::models::Link {
-                id: link.id.clone(),
-                org_id: link.org_id.clone(),
-                short_code: link.short_code.clone(),
-                destination_url: link.destination_url.clone(),
-                title: link.title.clone(),
-                created_by: link.created_by.clone(),
-                created_at: link.created_at,
-                updated_at: link.updated_at,
-                expires_at: link.expires_at,
-                status: crate::models::link::LinkStatus::Active,
-                click_count: link.click_count,
-                tags: link.tags, // Handle missing tags field
-                utm_params: link.utm_params,
-                forward_query_params: link.forward_query_params,
-                redirect_type: link.redirect_type.clone(),
-            };
-
-            // Resolve forward query params (same logic as in other places)
-            let resolved_forward = if let Some(forward) = link.forward_query_params {
-                forward
-            } else {
-                db::get_org_forward_query_params(&db, &link.org_id)
-                    .await
-                    .unwrap_or(false)
-            };
-
-            let mapping = link_model.to_mapping(resolved_forward);
-            kv::store_link_mapping(&kv, &link.org_id, &link.short_code, &mapping).await?;
-        }
-        "blocked" | "disabled" => {
-            // Link should NOT be active in KV - delete the KV entry
-            kv::delete_link_mapping(&kv, &link.org_id, &link.short_code).await?;
-        }
-        _ => {
-            return Response::error("Cannot sync link with unknown status", 400);
-        }
-    }
-
-    Response::from_json(&serde_json::json!({
-        "success": true,
-        "message": "Link KV entry re-synced successfully"
-    }))
-}
-
-// Blacklist handlers moved to api/admin/blacklist.rs
-
-// handle_admin_suspend_user, handle_admin_unsuspend_user, handle_admin_delete_user,
-// handle_admin_update_user_role moved to api/admin/users.rs
 
 #[utoipa::path(
     get,
